+++ libraries/DS2438/CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-01-27
+- add **float readVAD()** and **float getVAD()**
+- add **uint32_t readDisconnectTime()**
+- add **uint32_t readEndOfChargeTime()**
+- add **void enableCCA()** and **void disableCCA()**
+- add **float readCCA()** and **float readDCA()**
+- add **float readRemaining()**
+- rename some functions
+- update readme.md
+- implement some optimizations
+- minor edits
+
+
 ## [0.1.0] - 2024-01-26
 - initial release (experimental, not tested with hardware)